reorganize components and utilities
