+++ routes/cats.ts
@@ -1,156 +1,23 @@
 import { Router, Request, Response } from 'express';
+import { calculateAge } from '../src/utils/date.js';
+import { renderNavigationItem, NavigationItemProps } from '../src/components/navigation.js';
+import { renderAboutSection } from '../src/components/about.js';
+import { catsData } from '../src/data/cats.js';
 
 const router: Router = Router();
-interface Age {
-    years: number;
-    months: number;
-}
-
-interface Cat {
-    alias: string;
-    displayName: string;
-    displayPronouns: string;
-    aboutItems: Array<string>;
-    navigationThumbnailImgSrc: string;
-    navigationThumbnailImgAlt: string;
-    coverPhotoImgSrc: string;
-    coverPhotoImgAlt: string;
-}
-
-interface NavigationItem {
-    targetAlias: string;
-    imgSrc: string;
-    imgAlt: string;
-    label: string;
-}
-
-const catsData: { [alias: string]: Cat } = {
-    groucho: {
-        alias: 'groucho',
-        displayName: 'Groucho',
-        displayPronouns: 'he/him',
-        aboutItems: [
-            'Conqueror of boxes',
-            'Will definitely let you know when he is hungry',
-            'Very food-motivated',
-            'Definitely need to hide food from him',
-            'First to run away and hide from new people',
-            'Likes to bother Chica (and me)',
-            'Cuddly and likes to lay on my lap',
-            'Chipped his right fang a long time ago',
-            'Loves climbing',
-            'Loves head scratches',
-            'Prefers horizontal scratch surfaces'
-        ],
-        coverPhotoImgSrc: '/img/about_groucho_1.jpg',
-        coverPhotoImgAlt: 'Groucho the Cat, laying down',
-        navigationThumbnailImgSrc: '/img/about_groucho_thumbnail.jpg',
-        navigationThumbnailImgAlt: 'Groucho the Cat, laying on a woven blanket'
-    },
-    chica: {
-        alias: 'chica',
-        displayName: 'Chica',
-        displayPronouns: 'she/her',
-        aboutItems: [
-            'The mess maker',
-            'Destroyer of furniture and decorations',
-            'Sheds A LOT',
-            'Very quiet',
-            'First to greet me when I come back home (often waiting by the door)',
-            'Very playful (loves chasing, especially loves toy springs)',
-            'Will growl at reflected light and bugs',
-            'Loves bird watching',
-            'Likes to sleep near me',
-            'Loves whole body pets',
-            'Prefers vertical scratch surfaces'
-        ],
-        coverPhotoImgSrc: '/img/about_chica_1.jpg',
-        coverPhotoImgAlt: 'Chica the Cat, laying down on a drying rack',
-        navigationThumbnailImgSrc: '/img/about_chica_thumbnail.jpg',
-        navigationThumbnailImgAlt: 'Chica the Cat, sitting by a window overlooking a city'
-    }
-};
-
-function calculateAge(birthDate: Date): Age {
-    const today = new Date();
-
-    let years = today.getFullYear() - birthDate.getFullYear();
-    let months = today.getMonth() - birthDate.getMonth();
-
-    if (months < 0 || (months === 0 && today.getDate() < birthDate.getDate())) {
-        years--;
-        months += 12;
-    }
-
-    if (today.getDate() < birthDate.getDate()) {
-        months--;
-    }
-
-    return { years, months };
-}
-
-function renderNavigationItem({ targetAlias, imgSrc, imgAlt, label }: NavigationItem): string {
-    // Create a user-friendly URL path - replace all underscores with hyphens
-    const urlPath = targetAlias.replace(/_/g, '-');
-
-    return /* html */ `
-        <div
-            class="group relative flex-1 cursor-pointer" 
-            hx-get="/api/cats/about/${targetAlias}" 
-            hx-target="#about-section" 
-            hx-swap="innerHTML show:window:top"
-            hx-push-url="/cats/${urlPath}"
-            hx-on::after-request="htmx.ajax('GET', '/api/cats/about_navigation/${targetAlias}', '#about-navigation-section')"
-        >
-            <img alt="${imgAlt}" src="${imgSrc}" class="max-h-[480px] w-full object-cover" />
-            <div class="backdrop-blur-0 absolute inset-0 flex items-center justify-center bg-black/0 transition-all duration-300 group-hover:bg-black/30 group-hover:backdrop-blur-xs">
-                <p class="font-family-playfair text-xl font-bold text-white opacity-0 transition-opacity duration-300 group-hover:opacity-100">
-                    ${label}
-                </p>
-            </div>
-        </div>
-    `;
-}
-
-interface AboutSectionProps {
-    imageSrc: string;
-    imageAlt: string;
-    title: string;
-    subtitle?: string;
-    aboutItems: Array<string>;
-}
-
-function renderAboutSection({ imageSrc, imageAlt, title, subtitle, aboutItems }: AboutSectionProps): string {
-    const listHtml = aboutItems.map((item) => `<li>${item}</li>`).join('');
-
-    return /* html */ `
-        <img
-            alt="${imageAlt}"
-            src="${imageSrc}"
-            class="max-h-[480px] w-full object-cover"
-        />
-        <div class="p-8">
-            <h2 class="${subtitle ? 'mb-2' : 'mb-4'} text-2xl font-bold text-gray-800">${title}</h2>
-            ${subtitle ? /* html */ `<p class="mb-4 text-gray-500 italic">${subtitle}</p>` : ''}
-            <ul role="list" class="list-inside list-disc space-y-2 text-gray-700">
-                ${listHtml}
-            </ul>
-        </div>
-    `;
-}
 
 router.get('/about_navigation/:view', (req: Request, res: Response): void => {
-    const view = req.params.view;
+    const { view } = req.params;
 
     const { groucho, chica } = catsData;
-    const grouchoAndChicaNavigationItem: NavigationItem = {
+    const grouchoAndChicaNavigationItem: NavigationItemProps = {
         targetAlias: 'groucho_and_chica',
         imgSrc: '/img/about_grouchi_thumbnail.jpg',
         imgAlt: 'Groucho and Chica together',
         label: 'Groucho & Chica'
     };
 
-    const navigationConfigs: Record<string, NavigationItem[]> = {
+    const navigationConfigs: Record<string, NavigationItemProps[]> = {
         groucho_and_chica: [
             {
                 targetAlias: groucho.alias,
