+++ src/utils/date.ts
@@ -1,0 +1,19 @@
+import { Age } from '../types/cat';
+
+export function calculateAge(birthDate: Date): Age {
+    const today = new Date();
+
+    let years = today.getFullYear() - birthDate.getFullYear();
+    let months = today.getMonth() - birthDate.getMonth();
+
+    if (months < 0 || (months === 0 && today.getDate() < birthDate.getDate())) {
+        years--;
+        months += 12;
+    }
+
+    if (today.getDate() < birthDate.getDate()) {
+        months--;
+    }
+
+    return { years, months };
+}
